refactor: convert account/page.tsx to async server component (#345)
Extract client-side data fetching into server-side provider and
split interactive forms into profile-form.tsx and password-form.tsx.

- Add AccountViewModel and getAccountViewModel provider
- Keep layout and hasPassword conditional on server side
- Extract ProfileForm and PasswordForm as client components
