+++ server/presentation/view-models/account.ts
@@ -1,0 +1,5 @@
+export type AccountViewModel = {
+  name: string;
+  email: string;
+  hasPassword: boolean;
+};
